Fix getMediaId() to retrieve the ID from URL (#153)
A new Crunchyroll update made it possible for https://www.crunchyroll.com/es/that-time-i-got-reincarnated-as-a-slime/episode-17-the-gathering-780734 as a URL. The current regex used to extract the media ID didn't work with the `/es` part. This change relaxes the regex a bit to allow the new type of URL to also work.